Add delay after midnight prior to script checkin.
Ran into an odd issue. Originally, I saw that taking away a delay prior to running the script worked fine and would actually stop the auto DnD turn on to not take effect at all. I think it might be a race condition where the DnD state on the system level isn’t getting updated prior to my script getting checked, so the app just thinks that the state is still the same and will not change the DnD. Adding a 0.2s delay prior to checking the state fixes this(from my tests, might have to extend this more if I run into cases where it’s still not updating), but unfortunately the Notification Center icon on the menubar flashes at midnight.